branch_service/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
Flask-Migrate==4.1.0
PyMySQL==1.1.0
requests==2.32.3
